+++ src/main/java/com/fastcampus/projectboardadmin/domain/AdminAccount.java
@@ -1,0 +1,106 @@
+package com.fastcampus.projectboardadmin.domain;
+
+import com.fastcampus.projectboardadmin.domain.constant.RoleType;
+import com.fastcampus.projectboardadmin.domain.converter.RoleTypesConverter;
+
+import java.util.Collection;
+import java.util.LinkedHashSet;
+import java.util.Objects;
+import java.util.Set;
+
+import javax.persistence.Column;
+import javax.persistence.Convert;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Index;
+import javax.persistence.Table;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@ToString(callSuper = true)
+@Table(indexes = {
+	@Index(columnList = "email", unique = true),
+	@Index(columnList = "createdAt"),
+	@Index(columnList = "createdBy")
+})
+@Entity
+public class AdminAccount extends AuditingFields {
+
+	@Id
+	@Column(length = 50)
+	private String userId;
+
+	@Setter
+	@Column(nullable = false)
+	private String userPassword;
+
+	@Convert(converter = RoleTypesConverter.class)
+	@Column(nullable = false)
+	private Set<RoleType> roleTypes = new LinkedHashSet<>();
+	@Setter
+	@Column(length = 100)
+	private String email;
+	@Setter
+	@Column(length = 100)
+	private String nickname;
+	@Setter
+	private String memo;
+
+	protected AdminAccount() {
+	}
+
+	private AdminAccount(String userId, String userPassword, Set<RoleType> roleTypes, String email,
+		String nickname,
+		String memo, String createdBy) {
+		this.userId = userId;
+		this.userPassword = userPassword;
+		this.roleTypes = roleTypes;
+		this.email = email;
+		this.nickname = nickname;
+		this.memo = memo;
+		this.createdBy = createdBy;
+		this.modifiedBy = createdBy;
+	}
+
+	public static AdminAccount of(String userId, String userPassword, Set<RoleType> roleTypes,
+		String email, String nickname, String memo) {
+		return AdminAccount.of(userId, userPassword, roleTypes, email, nickname, memo, null);
+	}
+
+	public static AdminAccount of(String userId, String userPassword, Set<RoleType> roleTypes,
+		String email, String nickname, String memo, String createdBy) {
+		return new AdminAccount(userId, userPassword, roleTypes, email, nickname, memo, createdBy);
+	}
+
+	public void addRoleType(RoleType roleType) {
+		this.getRoleTypes().add(roleType);
+	}
+
+	public void addRoleTypes(Collection<RoleType> roleTypes) {
+		this.getRoleTypes().addAll(roleTypes);
+	}
+
+	public void removeRoleType(RoleType roleType) {
+		this.getRoleTypes().remove(roleType);
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o) {
+			return true;
+		}
+		if (!(o instanceof AdminAccount that)) {
+			return false;
+		}
+		return this.getUserId() != null && this.getUserId().equals(that.userId);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(this.getUserId());
+	}
+
+}
