+++ src/main/java/com/fastcampus/projectboardadmin/dto/UserAccountDto.java
@@ -1,63 +1,37 @@
 package com.fastcampus.projectboardadmin.dto;
 
-import com.fastcampus.projectboardadmin.domain.UserAccount;
+import com.fastcampus.projectboardadmin.domain.AdminAccount;
 import com.fastcampus.projectboardadmin.domain.constant.RoleType;
+
 import java.time.LocalDateTime;
 import java.util.Set;
 
 /**
- * A DTO for the {@link com.fastcampus.projectboardadmin.domain.UserAccount} entity
+ * A DTO for the {@link AdminAccount} entity
  */
 public record UserAccountDto(
-    String userId,
-    String userPassword,
-    Set<RoleType> roleTypes,
-    String email,
-    String nickname,
-    String memo,
-    LocalDateTime createdAt,
-    String createdBy,
-    LocalDateTime modifiedAt,
-    String modifiedBy
+	String userId,
+	Set<RoleType> roleTypes,
+	String email,
+	String nickname,
+	String memo,
+	LocalDateTime createdAt,
+	String createdBy,
+	LocalDateTime modifiedAt,
+	String modifiedBy
 ) {
 
-    public static UserAccountDto of(String userId, String userPassword, Set<RoleType> roleTypes,
-        String email, String nickname, String memo) {
-        return UserAccountDto.of(userId, userPassword, roleTypes, email, nickname, memo, null, null,
-            null, null);
-    }
+	public static UserAccountDto of(String userId, Set<RoleType> roleTypes,
+		String email, String nickname, String memo) {
+		return UserAccountDto.of(userId, roleTypes, email, nickname, memo, null, null,
+			null, null);
+	}
 
-    public static UserAccountDto of(String userId, String userPassword, Set<RoleType> roleTypes,
-        String email, String nickname, String memo, LocalDateTime createdAt, String createdBy,
-        LocalDateTime modifiedAt, String modifiedBy) {
-        return new UserAccountDto(userId, userPassword, roleTypes, email, nickname, memo, createdAt,
-            createdBy, modifiedAt, modifiedBy);
-    }
-
-    public static UserAccountDto from(UserAccount entity) {
-        return new UserAccountDto(
-            entity.getUserId(),
-            entity.getUserPassword(),
-            entity.getRoleTypes(),
-            entity.getEmail(),
-            entity.getNickname(),
-            entity.getMemo(),
-            entity.getCreatedAt(),
-            entity.getCreatedBy(),
-            entity.getModifiedAt(),
-            entity.getModifiedBy()
-        );
-    }
-
-    public UserAccount toEntity() {
-        return UserAccount.of(
-            userId,
-            userPassword,
-            roleTypes,
-            email,
-            nickname,
-            memo
-        );
-    }
+	public static UserAccountDto of(String userId, Set<RoleType> roleTypes,
+		String email, String nickname, String memo, LocalDateTime createdAt, String createdBy,
+		LocalDateTime modifiedAt, String modifiedBy) {
+		return new UserAccountDto(userId, roleTypes, email, nickname, memo, createdAt,
+			createdBy, modifiedAt, modifiedBy);
+	}
 
 }
