+++ src/test/java/com/fastcampus/projectboardadmin/controller/ArticleCommentManagementControllerTest.java
@@ -106,7 +106,6 @@ class ArticleCommentManagementControllerTest {
 	private UserAccountDto createUserAccountDto() {
 		return UserAccountDto.of(
 			"unoTest",
-			"pw",
 			Set.of(RoleType.ADMIN),
 			"uno-test@email.com",
 			"uno-test",
