+++ src/test/java/com/fastcampus/projectboardadmin/service/ArticleCommentManagementServiceTest.java
@@ -50,7 +50,6 @@ class ArticleCommentManagementServiceTest {
 	private UserAccountDto createUserAccountDto() {
 		return UserAccountDto.of(
 			"unoTest",
-			"pw",
 			Set.of(RoleType.ADMIN),
 			"uno-test@email.com",
 			"uno-test",
