+++ src/test/java/com/fastcampus/projectboardadmin/service/ArticleManagementServiceTest.java
@@ -164,7 +164,6 @@ class ArticleManagementServiceTest {
 		private UserAccountDto createUserAccountDto() {
 			return UserAccountDto.of(
 				"unoTest",
-				"pw",
 				Set.of(RoleType.ADMIN),
 				"uno-test@email.com",
 				"uno-test",
